requirements.txt
Django~=3.1.7
channels~=3.0.3
djangorestframework~=3.12.4
celery~=5.0.5
redis~=3.5.3
requests~=2.25.1
asgiref~=3.3.1
pytickersymbols~=1.4.3
pandas~=1.2.3
django-cors-headers~=3.7.0
channels-redis~=3.2.0
pytest=1.8.2